Refactors helpers and types which are used in multiple places to live in their respective packages
