+++ packages/custom-types/index.d.ts
@@ -8,11 +8,14 @@ export { default as Channel } from './channel'
 export { default as ForceData } from './force-data'
 export { default as Force } from './force'
 export { default as Message } from './message'
+export { default as MappingContext} from './mapping-context'
 export { default as Participant } from './participant'
 export { default as PlatformTypeData } from './platform-type-data'
 export { default as PlatformType } from './platform-type'
 export { default as Role } from './role'
 export { default as States } from './states'
+export { default as SergeHex } from './serge-hex'
+export { default as SergeGrid } from './serge-grid'
 export { default as Template } from './template'
 export { default as WargameData } from './wargame-data'
 export { default as WargameList } from './wargame-list'
