+++ packages/helpers/allowable-cells.ts
@@ -1,5 +1,4 @@
-import SergeHex from '../types/serge-hex'
-import SergeGrid from '../types/serge-grid'
+import { SergeHex, SergeGrid } from '@serge/custom-types'
 
 /**
  *  create hexagonal grid
