+++ packages/helpers/index.ts
@@ -3,7 +3,11 @@
 * To generate an updated version, please run `yarn genreg`
 ***/
 
-export { default as findPerceivedAsTypes } from './findPerceivedAsTypes'
+export { default as calcAllowableCells } from './allowable-cells'
 export { default as getSelectedItem } from './getSelectedItem'
+export { default as hexNamed } from './hex-named'
 export { default as padInteger } from './padInteger'
+export { default as plannedRouteFor } from './planned-route-for'
 export { default as roundToNearest } from './roundToNearest'
+export { default as toScreen } from './to-screen'
+export { default as toWorld } from './to-world'
