+++ packages/helpers/package.json
@@ -3,5 +3,8 @@
   "version": "0.3.0",
   "license": "Apache-2.0",
   "author": "foxleigh81 <alex@alexfoxleigh.com>",
-  "main": "build/index.js"
+  "main": "build/index.js",
+  "dependencies": {
+    "honeycomb-grid": "^3.1.7"
+  }
 }
